refactor: entrypoints
