+++ static/index.html
@@ -2,13 +2,15 @@
 <html>
   <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
-    <meta name="viewport" content="initial-scale=1, maximum-scale=1" />
+
     <style>
       html,
       body {
         margin: 0;
         padding: 0;
         height: 100%;
+        overflow: hidden;
+        touch-action: none;
       }
 
       * {
@@ -20,28 +22,24 @@
         color: #333;
         font-family: 'open sans', roboto, 'helvetica neue', sans-serif;
         font-size: 0.8em;
-        overflow: hidden;
       }
 
-      div.loading-image {
-        display: none;
+      .dcl-loading {
+        background: url(images/decentraland-loading.gif) center center no-repeat;
       }
 
-      .dcl-loading div.loading-image {
+      canvas {
+        position: relative;
+        z-index: 1000;
         width: 100%;
         height: 100%;
-        background: url(images/decentraland-loading.gif) center center no-repeat white;
-        display: block;
-      }
-
-      .dcl-loading .minimap {
-        display: none;
       }
     </style>
+    <script src="https://code.jquery.com/pep/0.4.3/pep.js"></script>
   </head>
 
   <body class="dcl-loading">
-    <div class="loading-image"></div>
+    <div id="gameContainer" style="width: 100%; height: 100%; margin: auto"></div>
     <script async src="https://www.googletagmanager.com/gtag/js?id=UA-63691338-5"></script>
     <script>
       window.dataLayer = window.dataLayer || []
@@ -52,10 +50,12 @@
       gtag('js', new Date())
       gtag('config', 'UA-63691338-5')
     </script>
+
     <script>
       // prettier-ignore
-      !function () { var analytics = window.analytics = window.analytics || []; if (!analytics.initialize) if (analytics.invoked) window.console && console.error && console.error("Segment snippet included twice."); else { analytics.invoked = !0; analytics.methods = ["trackSubmit", "trackClick", "trackLink", "trackForm", "pageview", "identify", "reset", "group", "track", "ready", "alias", "debug", "page", "once", "off", "on"]; analytics.factory = function (t) { return function () { var e = Array.prototype.slice.call(arguments); e.unshift(t); analytics.push(e); return analytics } }; for (var t = 0; t < analytics.methods.length; t++) { var e = analytics.methods[t]; analytics[e] = analytics.factory(e) } analytics.load = function (t, e) { var n = document.createElement("script"); n.type = "text/javascript"; n.async = !0; n.src = "https://cdn.segment.com/analytics.js/v1/" + t + "/analytics.min.js"; var a = document.getElementsByTagName("script")[0]; a.parentNode.insertBefore(n, a); analytics._loadOptions = e }; analytics.SNIPPET_VERSION = "4.1.0"; } }();
+      !function(){var analytics=window.analytics=window.analytics||[];if(!analytics.initialize)if(analytics.invoked)window.console&&console.error&&console.error("Segment snippet included twice.");else{analytics.invoked=!0;analytics.methods=["trackSubmit","trackClick","trackLink","trackForm","pageview","identify","reset","group","track","ready","alias","debug","page","once","off","on"];analytics.factory=function(t){return function(){var e=Array.prototype.slice.call(arguments);e.unshift(t);analytics.push(e);return analytics}};for(var t=0;t<analytics.methods.length;t++){var e=analytics.methods[t];analytics[e]=analytics.factory(e)}analytics.load=function(t,e){var n=document.createElement("script");n.type="text/javascript";n.async=!0;n.src="https://cdn.segment.com/analytics.js/v1/"+t+"/analytics.min.js";var a=document.getElementsByTagName("script")[0];a.parentNode.insertBefore(n,a);analytics._loadOptions=e};analytics.SNIPPET_VERSION="4.1.0";}}();
     </script>
-    <script defer async src="dist/index.js"></script>
   </body>
+  <script src="unity/Build/UnityLoader.js"></script>
+  <script defer async src="dist/index.js"></script>
 </html>
